libra: mount "cust" partition to /cust by default
Just found there are 301MB free space in "cust" partition of Mi-4c, which could be made use of in the future.
The good thing is the "vendor" part of the system image is 305MB, which is less than the free space in "cust".
So a (fake) project treble might be helpful to solve the "too little free space" problem of "system" partition, without modifying the partition table.