libra: Enable zram for both 2GB/3GB devices
ZRAM improves the user experience on low-ram devices, but may decreace the performance if the device has enough RAM.
Thanks to the "ZSTD" crypto algorithm, the performance of ZRAM should be much better now.
According to my tests, ZRAM is also seldom used on 3GB devices, so the impact will be even less.