libra: Add necessary manual comments
[NOTE 1]
The kernel is now available to build with GCC 10.1 on Ubuntu 20.10 (or 20.04).
Using GCC 9 or lower versions is still possible, but has no/less benefit on performance.
https://github.com/WJXXBSH/android_kernel_xiaomi_libra

[NOTE 2]
To build all apps in the rom with "speed" filter (for better performance), please refer to the diff of "target/product/runtime_libart.mk" on "project build/make":
https://github.com/WJXXBSH/useful_android_scripts/blob/master/repo/repo_16.diff