libra: Add new property for edge gesture
It's now the time to separate JAVA file operation out of LineageParts, for better performance and stability.
If "persist.vendor.edge_touch_mode" is set to true, init will write "1" to "/proc/touchscreen/edge_touch_mode" (edge gesture enabled), otherwise "0" (edge gesture disabled).

As "edge_touch_mode" is default to "0", it's not necessary to set "0" again if edge gesture is disabled.
There will be a toggle in Settings app to control edge gesture status easily (for my LineageOS roms), which needs both "false" and "true" for "persist.vendor.edge_touch_mode" to be handled.

All patches are removed in this commit as they're not generated from lineage-16.0 sources.