libra: Build kernel image with clang-12
It's now possible to install clang-12 package on both Debian and Ubuntu.

VDSO32 must be built with arm toolchains (instead of arm64), but it's not possible to specify custom arm toolchains when building an arm64 kernel.
As a result, the prebuilt gcc-4.9 toolchains will be used unconditionally, even if "CONFIG_CROSS_COMPILE_ARM32" is set in kernel config.
To resolve the trouble caused by VDSO32 in kernel, I need to modify "vendor/lineage/config/BoardConfigKernel.mk" and set "KERNEL_TOOLCHAIN_*" to proper values directly.