Source code clean-up [#5]
A merged diff file will be included in this repo instead of git patches, which might be unable to apply (using "git am xxx.patch") after some upstream changes.
Also remove some scripts that are not recommended to use (the repo "android_vendor_xiaomi_libra" should be a better choice).

Note: the "#5" in commit message means "the 5th time of clean-up", but not "issue/pull No.5 of this repo".